requirement.txt
Flask
psycopg2
Blueprint
python-dotenv